Move Kotlin DSL tests from integ-test subproject
